Update paginator.php
Fixed the problem when the page has no results. Just have to check on Front where you have to take the first position if is null or not, anyway the second index position give you always the last page. PD: On Front before render the php file, if the current page where we are, is more than the last one we got, just redirect to the last page like this: 

$lastPage = $paginationData[1];//LAST PAGE VARIABLE
if($currentPage>$lastPage){
    header ("Location: ?page=".$lastPage);
}